+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 
 ### vNEXT
 
-- **core**: BREAKING - `diff` is now async
+- **core**: BREAKING CHANGE - `diff` is now async
+- **github**: BREAKING CHANGE - `experimental_merge` is now enabled by default
 - **core**: Adds `considerUsage` rule
 - **core**: Fixes missing names of default root types
+- **cli**, **ci**: Adds `@aws_lambda` directive
+- **cli**, **ci**: Fixes missing headers in diff command
 
 ### v2.9.0